arreglados los fallos en MMM

+++ Proyectos/MMM/src/interfazDeUsuario/Calculable.java
@@ -1,6 +1,6 @@
 package interfazDeUsuario;
 
 public interface Calculable {
-    int getCalcularPrecio();
-    
-} 
+    float getCalcularPrecio();
+
+}
